+++ db/migrate/20260530120000_default_notify_via_slack_to_true.rb
@@ -1,0 +1,5 @@
+class DefaultNotifyViaSlackToTrue < ActiveRecord::Migration[8.1]
+  def change
+    change_column_default :users, :notify_via_slack, from: false, to: true
+  end
+end