+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_05_29_100000) do
+ActiveRecord::Schema[8.1].define(version: 2026_05_30_120000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -187,7 +187,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_29_100000) do
     t.string "email", null: false
     t.string "name"
     t.boolean "notify_via_email", default: true, null: false
-    t.boolean "notify_via_slack", default: false, null: false
+    t.boolean "notify_via_slack", default: true, null: false
     t.string "provider", null: false
     t.string "slack_handle"
     t.string "slack_id"